refactor(Navigation/Table): move DownloadManager to base components [YTFRONT-5970]
