+++ packages/ui/src/ui/components/DownloadManager/DownloadManager.scss
