+++ packages/ui/src/ui/components/DownloadManager/SeparatorInput.scss
