+++ packages/ui/src/ui/components/DownloadManager/i18n/dicts.ts
