+++ packages/ui/src/ui/components/DownloadManager/i18n/index.ts
@@ -1,4 +1,4 @@
-import {addI18Keysets} from '../../../../../../i18n';
+import {addI18Keysets} from '../../../i18n';
 
 import dicts from './dicts';
