+++ packages/ui/src/ui/components/DownloadManager/index.ts
@@ -1,0 +1,1 @@
+export {DownloadManager} from './DownloadManager';
